+++ CHANGELOG.md
@@ -13,6 +13,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   ([#1619](https://github.com/feldera/feldera/pull/1619))
 - SQL: Support for `ARRAYS_OVERLAP` function ([#1570](https://github.com/feldera/feldera/pull/1570))
 
+### Changed
+
+- [SQL]: `ARRAY_CONTAINS`, `ARRAY_POSITION`, `ARRAY_REMOVE` functions
+  for arguments of different types ([#1636](https://github.com/feldera/feldera/pull/1636))
+
 ## [0.13.0] - 2024-04-02
 
 ### Added